+++ CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New configuration field `window.position` allows specifying the starting position
 - Added the ability to change the selection color
 - Text will reflow instead of truncating when resizing Alacritty
+- Underline text and change cursor when hovering over URLs with required modifiers pressed
+
+### Changed
+
+- Clicking on non-alphabetical characters in front of URLs will no longer open them
 
 ### Fixed